requirements.txt
# requirements.txt
transformers
sentence-transformers
faiss-cpu
PyMuPDF
gradio
numpy
torch